# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /router/status:
    get:
      operationId: getRouterStatus
      tags: [router]
      security: [{ bearerAuth: [] }]
      summary: Get Router IA configuration and connection status
      responses:
        "200":
          description: Router IA configuration state; secret values are never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouterConnectionStatus"
  /router/providers:
    get:
      operationId: listRouterProviders
      tags: [router]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: User-owned providers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RouterProvider" }
    post:
      operationId: createRouterProvider
      tags: [router]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RouterProviderInput" }
      responses:
        "201":
          description: Provider created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RouterProvider" }
  /router/providers/{providerId}:
    parameters:
      - name: providerId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateRouterProvider
      tags: [router]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RouterProviderUpdate" }
      responses:
        "200":
          description: Provider updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RouterProvider" }
    delete:
      operationId: deleteRouterProvider
      tags: [router]
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Provider deleted }
  /router/providers/{providerId}/test:
    parameters:
      - name: providerId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: testRouterProvider
      tags: [router]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Provider test result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RouterProviderTestResult" }
  /router/test:
    post:
      operationId: testRouterConnection
      tags: [router]
      security: [{ bearerAuth: [] }]
      summary: Test the Router IA connection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouterConnectionTestInput"
      responses:
        "200":
          description: Connection status after a real Router IA request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouterConnectionStatus"
        "400":
          description: Invalid Router test request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    RouterTaskType:
      type: string
      enum: [chat, coding, reasoning, summarization, vision, document]
    RouterConnectionStatus:
      type: object
      properties:
        configured:
          type: boolean
        connected:
          type: boolean
        lastTestAt:
          type: ["string", "null"]
          format: date-time
        lastLatencyMs:
          type: ["integer", "null"]
          minimum: 0
        message:
          type: ["string", "null"]
          maxLength: 240
      required:
        - configured
        - connected
        - lastTestAt
        - lastLatencyMs
        - message
    RouterConnectionTestInput:
      type: object
      properties:
        task_type:
          type: string
          enum: [chat]
      required: [task_type]
    RouterProvider:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string, enum: [openai, anthropic, gemini, openai-compatible, groq] }
        baseUrl: { type: ["string", "null"] }
        model: { type: string }
        apiKeyPreview: { type: string }
        isDefault: { type: boolean }
        capabilities: { type: array, minItems: 1, items: { $ref: "#/components/schemas/RouterTaskType" } }
        priority: { type: integer }
        isActive: { type: boolean }
        status: { type: string }
        healthStatus: { type: string }
        lastTestAt: { type: ["string", "null"], format: date-time }
      required: [id, name, kind, baseUrl, model, apiKeyPreview, isDefault, capabilities, priority, isActive, status, healthStatus, lastTestAt]
    RouterProviderInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        kind: { type: string, enum: [openai, anthropic, gemini, openai-compatible, groq] }
        baseUrl: { type: ["string", "null"], maxLength: 2048 }
        model: { type: string, minLength: 1, maxLength: 160 }
        apiKey: { type: string, minLength: 1, maxLength: 4096 }
        isDefault: { type: boolean }
        capabilities: { type: array, minItems: 1, items: { $ref: "#/components/schemas/RouterTaskType" } }
        priority: { type: integer, minimum: 0, maximum: 1000 }
        isActive: { type: boolean }
      required: [name, kind, model, apiKey]
    RouterProviderUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        baseUrl: { type: ["string", "null"], maxLength: 2048 }
        model: { type: string, minLength: 1, maxLength: 160 }
        apiKey: { type: string, minLength: 1, maxLength: 4096 }
        isDefault: { type: boolean }
        capabilities: { type: array, minItems: 1, items: { $ref: "#/components/schemas/RouterTaskType" } }
        priority: { type: integer, minimum: 0, maximum: 1000 }
        isActive: { type: boolean }
    RouterProviderTestResult:
      type: object
      properties:
        ok: { type: boolean }
        latencyMs: { type: ["integer", "null"] }
        message: { type: string }
      required: [ok, latencyMs, message]
    ApiError:
      type: object
      properties:
        error:
          type: string
          maxLength: 240
      required: [error]