# lib/api-zod/src/generated/types/routerProvider.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RouterProviderKind } from './routerProviderKind';
import type { RouterTaskType } from './routerTaskType';

export interface RouterProvider {
  id: string;
  name: string;
  kind: RouterProviderKind;
  /** @nullable */
  baseUrl: string | null;
  model: string;
  apiKeyPreview: string;
  isDefault: boolean;
  /** @minItems 1 */
  capabilities: RouterTaskType[];
  priority: number;
  isActive: boolean;
  status: string;
  healthStatus: string;
  /** @nullable */
  lastTestAt: Date | null;
}