# lib/api-zod/src/generated/types/routerProviderUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RouterTaskType } from './routerTaskType';

export interface RouterProviderUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name?: string;
  /**
     * @maxLength 2048
     * @nullable
     */
  baseUrl?: string | null;
  /**
     * @minLength 1
     * @maxLength 160
     */
  model?: string;
  /**
     * @minLength 1
     * @maxLength 4096
     */
  apiKey?: string;
  isDefault?: boolean;
  /** @minItems 1 */
  capabilities?: RouterTaskType[];
  /**
     * @minimum 0
     * @maximum 1000
     */
  priority?: number;
  isActive?: boolean;
}